+++ helpers/menu_xxx.js
@@ -1,7 +1,7 @@
 ﻿'use strict';
 
 /* 
-	Contextual Menu helper v 1.1 19/03/21
+	Contextual Menu helper v 1.2 19/05/21
 	Helper to create contextual menus on demand on panels without needing to create specific methods for
 	every script, calculate IDs, etc. Menus are pushed to a list and created automatically, linking the entries
 	to their idx without needing a 'switch' block or leaving holes to ensure idx get enough numbers to expand the script.
@@ -13,9 +13,9 @@
 			-bSupressDefaultMenu:	Suppress the default context menu. left shift + left windows key will bypass it. 
 			-idxInitial:			Specifies an initial idx to create menus (useful to concatenate multiple menus objects)
 			
-		.btn_up(x, y, object)
-			-NOTE: 		Called within callbacks to create the menu. Specifying an object (like another menu instance), lets you
-						concatenate multiple menus. Uses object.btn_up() and object.btn_up_done()
+		.btn_up(x, y, [object])
+			-NOTE: 		Called within callbacks to create the menu. Specifying an object or array of objects (like another menu instances), 
+						lets you concatenate multiple menus. Uses object.btn_up() and object.btn_up_done() on manually added entries
 		.getMainMenuName()
 			-NOTE:		Used to get the key of the main menu. Useful to concatenate multiple menus.
 	
@@ -43,72 +43,16 @@
 			-entryText: Just for information
 			-NOTE:		Used to create dynamic menus only when calling the contextual menu, useful to check for tracks selection, etc.
 						You may use any other method like .newMenu(), .newEntry(), etc. within condFunc. Thus creating menus only if required.
-	
-	For example:
-		-Standard menu:
-			var menu = new _menu();
-			menu.newEntry({entryText: 'Hola', func: () => {console.log('hola')}});
-			menu.newEntry({entryText: 'sep'});
-			menu.newEntry({entryText: 'Hola2', func: () => {console.log('hola2')}});
-			var bSubMenu = true;
-			const funct = () => {return (bSubMenu) ? 'SubMenu 1' : 'SubMenu 2';};
-			menu.newMenu(funct);
-			menu.newEntry({menuName: funct, entryText: 'Change SubMenu', func: () => {bSubMenu = !bSubMenu}});
-			menu.newEntry({menuName: funct, entryText:'Hola 3', func: () => {console.log('hola3')}, flags: () => {return (bSubMenu) ? MF_STRING : MF_GRAYED}});
-			menu.newCheckMenu(funct, 'Change SubMenu', 'Hola 3', () => {return (bSubMenu) ? 0 : 1;});
-			menu.newEntry({entryText: 'Hola 4', func: () => {console.log('hola4')}});
-			
-			function on_mouse_rbtn_up(x, y) {return menu.btn_up(x, y);}
-		
-		Renders to: (note 'Hola 4' Entry is drawn after the sub-menu)
-			+Hola
-			+-----
-			+HOla 2
-			+SubMenu 1 / SubMenu 2:
-				+Change SubMenu
-				+Hola 3
-			+Hola 4
-		
-		- Manually adding some entries to the menu (uses previous code too):
-			var menuTwo = new _menuTwo();
-			function _menuTwo() {
-				this.idxInitial = 0;
-
-				this.btn_up = (idxInitial) => {
-					this.idxInitial = idxInitial;
-					const menuName = menu.getMainMenuName();
-					menu.getMenu(menuName).AppendMenuItem(MF_STRING, idxInitial + 1, 'Manual entry 1');
-				}
-				
-				this.btn_up_done = (currIdx) => {
-					if (currIdx == this.idxInitial + 1) {
-						console.log('Manual entry 1');
-						return;
-					} else {return;}
-				}
-			}
-			
-			function on_mouse_rbtn_up(x, y) {return menu.btn_up(x, y, menuTwo);}
-			
-		Renders to:
-			+Hola
-			+-----
-			+HOla 2
-			+SubMenu 1 / SubMenu 2:
-				+Change SubMenu
-				+Hola 3
-			+Hola 4
-			+Manual entry 1
-		
+	Usage:
+		See examples folder.
+						
 	TODO:
 		- Add invisible IDs to entries names (?)
-		- Add properties to the object to simplify its use
  */
 
 include(fb.ComponentPath + 'docs\\Flags.js');
-include(fb.ProfilePath + 'scripts\\SMP\\xxx-scripts\\helpers\\helpers_xxx.js');
 
-function _menu({bSupressDefaultMenu = true, idxInitial = 0} = {}) {
+function _menu({bSupressDefaultMenu = true, idxInitial = 0, properties = null} = {}) {
 	var menuArr = [];
 	var menuMap = new Map();
 	var entryArrTemp = [];
@@ -119,6 +63,8 @@ function _menu({bSupressDefaultMenu = true, idxInitial = 0} = {}) {
 	var checkMenuMap = new Map();
 	var checkMenuArr = [];
 	var idx = idxInitial;
+	
+	this.properties = properties; // To simplify usage along other scripts
 	
 	// To create new elements
 	this.newMenu = (menuName = 'main', subMenuFrom = 'main') => {
@@ -144,6 +90,7 @@ function _menu({bSupressDefaultMenu = true, idxInitial = 0} = {}) {
 	}
 
 	this.getNumEntries = () => {return entryArr.length;}
+	this.getEntries = () => {return [...entryArr];}
 	this.getMainMenuName = () => {return menuArr[0].menuName;}
 	this.hasMenu = (menuName) => {return (menuArr.indexOf(menuName) !== -1);}
 	
@@ -154,7 +101,7 @@ function _menu({bSupressDefaultMenu = true, idxInitial = 0} = {}) {
 	this.getCheckMenu = (menuName) => {return (!menuName) ? checkMenuMap : checkMenuMap.get(menuName);}
 	
 	this.createMenu = (menuName = menuArr[0].menuName) => {
-		if (_isFunc(menuName)) {menuName = menuName();}
+		if (_isFunction(menuName)) {menuName = menuName();}
 		menuMap.set(menuName, window.CreatePopupMenu());
 		return menuMap.get(menuName);
 	}
@@ -163,9 +110,9 @@ function _menu({bSupressDefaultMenu = true, idxInitial = 0} = {}) {
 		if (entryText === 'sep' || entryText === 'separator') {menuMap.get(menuName).AppendMenuSeparator();}
 		else {
 			idx++;
-			if (_isFunc(menuName)) {menuName = menuName();}
-			if (_isFunc(flags)) {flags = flags();}
-			if (_isFunc(entryText)) {entryText = entryText();}
+			if (_isFunction(menuName)) {menuName = menuName();}
+			if (_isFunction(flags)) {flags = flags();}
+			if (_isFunction(entryText)) {entryText = entryText();}
 			menuMap.get(menuName).AppendMenuItem(flags, idx, entryText);
 			entryMap.set(entryText, idx);
 			idxMap.set(idx, func);
@@ -174,16 +121,39 @@ function _menu({bSupressDefaultMenu = true, idxInitial = 0} = {}) {
 	
 	this.checkMenu = (menuName, entryTextA, entryTextB, idxFunc) => {
 		checkMenuMap.set(menuName, () => {
-			if (_isFunc(menuName)) {menuName = menuName();}
-			if (_isFunc(entryTextA)) {entryTextA = entryTextA();}
-			if (_isFunc(entryTextB)) {entryTextB = entryTextB();}
+			if (_isFunction(menuName)) {menuName = menuName();}
+			if (_isFunction(entryTextA)) {entryTextA = entryTextA();}
+			if (_isFunction(entryTextB)) {entryTextB = entryTextB();}
 			return menuMap.get(menuName).CheckMenuRadioItem(this.getIdx(entryTextA), this.getIdx(entryTextB), this.getIdx(entryTextA) + idxFunc());
 		});
+	}
+	
+	this.concat = (menuObj) => {
+		entryArr = entryArr.concat(menuObj.getEntries());
 	}
 	
 	this.btn_up = (x, y, object, forcedEntry = '') => {
 		// Add conditional entries/menus
 		entryArrTemp = [...entryArr]; // Create backup to restore later
+		// Call other object's menu creation. It allows multiple instances of this framework, either manually appending items
+		// or an entire new instance. Separators may be added too.
+		let objectArr = [];
+		let manualMenuArr = [];
+		if (object) {
+			if (isArray(object)) {objectArr = object;}
+			else {objectArr = [object];}
+			for (const objectMenu of objectArr) {
+				if (compareKeys(this, objectMenu)) { // Another instance of this framework, just merge entries and done
+					this.concat(objectMenu);
+				} else if (objectMenu === 'sep' || objectMenu === 'separator') { // Separator
+					this.newEntry({entryText: 'sep'});
+				} else if (objectMenu.hasOwnProperty('btn_up') && objectMenu.hasOwnProperty('btn_up_done')) { // Object with hard-coded methods
+					manualMenuArr.push(objectMenu);
+				} else { // Error
+					console.log('menu_xxx: Tried to merge an external menu withouth known methods (\'' + (typeof objectMenu === 'object' ? JSON.stringify(Object.keys(objectMenu)) : objectMenu) + '\').');
+				}
+			}
+		}
 		entryArr.forEach( (entry) => {
 			if (entry.hasOwnProperty('condFunc') && entry.condFunc) { // Create menu
 				entry.condFunc();
@@ -199,9 +169,9 @@ function _menu({bSupressDefaultMenu = true, idxInitial = 0} = {}) {
 			if (!entry.bIsMenu) { // To main menu
 				this.addToMenu({entryText: entry.entryText, func: entry.func, menuName: entry.menuName, flags: entry.flags});
 			} else { // Append sub-menus
-				const subMenuName = _isFunc(entry.menuName) ? entry.menuName() : entry.menuName;
+				const subMenuName = _isFunction(entry.menuName) ? entry.menuName() : entry.menuName;
 				if (subMenuName !== menuArr[0].menuName) {
-					const subMenuFrom = _isFunc(entry.subMenuFrom) ? entry.subMenuFrom() : entry.subMenuFrom;
+					const subMenuFrom = _isFunction(entry.subMenuFrom) ? entry.subMenuFrom() : entry.subMenuFrom;
 					this.getMenu(subMenuName).AppendTo(this.getMenu(subMenuFrom), MF_STRING, subMenuName);
 				}
 			}
@@ -213,9 +183,12 @@ function _menu({bSupressDefaultMenu = true, idxInitial = 0} = {}) {
 		this.getCheckMenu().forEach( (func) => {
 			func();
 		});
-		// Call other object's menu creation
-		if (object && object.hasOwnProperty('btn_up')) {
-			object.btn_up(this.getNumEntries()); // The current num of entries may be used to create another menu with that initial idx
+		// Call other object's menu creation manually appended items
+		if (manualMenuArr.length) {
+			let idxAcum = this.getNumEntries();
+			for (const objectMenu of manualMenuArr) {
+				idxAcum = objectMenu.btn_up(idxAcum); // The current num of entries may be used to create another menu with that initial idx, subsequent calls should return/use the new idx
+			}
 		}
 		// Find currently selected item
 		const currIdx = forcedEntry.length ? this.getIdx(forcedEntry) : this.getMenu(menuArr[0].menuName).TrackPopupMenu(x, y);
@@ -225,12 +198,15 @@ function _menu({bSupressDefaultMenu = true, idxInitial = 0} = {}) {
 				if (bDone) {return;}
 				if (entryIdx === currIdx) {
 					func();
-					return bDone = true;
+					bDone = true;
+					return;
 				}
 			});
 			// Call other object's menu selection
-			if (!bDone && object && object.hasOwnProperty('btn_up_done')) {
-				object.btn_up_done(currIdx);
+			if (!bDone && manualMenuArr.length) {
+				for (const objectMenu of manualMenuArr) {
+					objectMenu.btn_up_done(currIdx);
+				}
 			}
 		} else if (forcedEntry.length) {
 			console.log('menu_xxx: Tried to call a menu with forced entry (\'' + forcedEntry + '\') but it doesn\'t exist. It may point to a bug or error.');
@@ -250,7 +226,20 @@ function _menu({bSupressDefaultMenu = true, idxInitial = 0} = {}) {
 	}
 }
 
-// Helper
-function _isFunc(obj) {
+// Helpers
+function _isFunction(obj) {
   return !!(obj && obj.constructor && obj.call && obj.apply);
+}
+
+function compareKeys(a, b) {
+	const aKeys = Object.keys(a).sort();
+	const bKeys = Object.keys(b).sort();
+	return JSON.stringify(aKeys) === JSON.stringify(bKeys);
+}
+
+function isArray(checkKeys) {
+	if ( checkKeys === null || Object.prototype.toString.call(checkKeys) !== '[object Array]' || checkKeys.length === null || checkKeys.length === 0){
+		return false; //Array was null or not an array
+	}
+	return true;
 }